+++ application/languages/en/views.php
@@ -1882,7 +1882,7 @@ return [
     'Voir' => 'View',
     //COI
     "Votre réponse à bien été enregistrée." => "Your answer has been saved.",
-    "Conflit d'intérêts(CI)" => "Conflict Of Interest(COI)",
+    "Conflit d'intérêts (CI)" => "Conflict Of Interest (COI)",
     "Activer la déclaration CI" => "Enable declaring COI",
     "Le mode conflit d'intérêts (CI) aura les effets suivants : toutes les informations non publiques concernant une soumission ne sont pas accessibles à un éditeur tant qu'il n'a pas déclaré l'absence de tout conflit d'intérêts" => "This mode will have the following effects: all non-public information about a submission is not accessible to an editor until he/she has declared the absence of any conflict of interest",
     "Vous avez été redirigé, car vous avez déclaré un conflit d'intérêts avec cette soumission . " => "You've been redirected, because you have declared a conflict of interest with this submission.",